\name{finland}
\alias{finland}
\docType{data}
\encoding{latin1}
\title{Data set for Finland, Johansen \& Juseliues (1990)}
\description{
  This data set contains the series used by S. Johansen and K. Juselius
  for estimating a money demand function of Finland.
}
\usage{data(finland)}
\format{
  A data frame with 106 observations on the following 4 variables,
  ranging from 1958:Q2 until 1984:Q3.

    \tabular{rl}{
    \code{lrm1} \tab Logarithm of real money, M1. \cr
    \code{lny} \tab Logarithm of real income. \cr
    \code{lnmr} \tab Marginal rate of interest. \cr
    \code{difp} \tab Inflation rate. \cr
    }
}
\source{
  Johansen, S. and Juselius, K. (1990), Maximum Likelihood Estimation and
  Inference on Cointegration -- with Applications to the Demand for
  Money, \emph{Oxford Bulletin of Economics and Statistics}, \bold{52,
    2}, 169--210.
  
}
\references{
  \url{http://www.math.ku.dk/~sjo/}
}
\author{Bernhard Pfaff}
\keyword{datasets}
