\name{Raotbl2} 
\alias{Raotbl2}
\docType{data}
\encoding{latin1}
\title{Data set used by Dickey, Jansen \& Thornton (1994)}
\description{
  This data set contains the time series used by David A. Dickey, Dennis
  W. Jansen and Daniel L. Thornton in their article: ``A Primer on
  Cointegrating with an Application to Money and Income''. 
}
\usage{data(Raotbl2)}
\format{
  A data frame with quarterly oberservations (ts objects) starting in
  1953:1 until 1988:4 for the following 4 variables (all transformed to
  natural logarithms.
  
    \tabular{rl}{
    \code{m1p} \tab Real money balances M1. \cr
    \code{m2p} \tab Real money balances M2. \cr
    \code{mbp} \tab Real adjusted monetary base. \cr
    \code{nm1m2p} \tab Real non-M1 component of M2.
    }
}
\source{
  Dickey, David A., Dennis W. Jansen and Daniel L. Thornton (1994), A
  Primer on Cointegration with an Application to Money and Income, in:
  Cointegration for the Applied Economist, ed. B. Bhaskara Rao, chapter
  2, Data Appendix, Table D.2.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Rao Dickey Jansen Thornton}
