\name{Raotbl3} 
\alias{Raotbl3}
\encoding{latin1}
\docType{data}
\title{Data set used by Holden and Perman (1994)}
\description{
  This data set contains the time series used by Darryl Holden and Roger
  Perman in their article: ``Unit Roots and Cointegration for the Economist". 
}
\usage{data(Raotbl3)}
\format{
  A data frame with quarterly data (ts objects) from the United Kingdom
  starting in 1966:4 until 1991:2 for the following 6 variables (all
  transformed to natural logarithms).
  
    \tabular{rl}{
    \code{lc} \tab Real consumption expenditure. \cr
    \code{li} \tab Real income. \cr
    \code{lw} \tab Real wealth. \cr
    \code{dd682} \tab Dummy variable for 68:2. \cr
    \code{dd792} \tab Dummy variable for 79:2. \cr
    \code{dd883} \tab Dummy variable for 88:3. 
  }
  
  More details about the data are provided in the data appendix of Rao,
  ``Cointegration for the Applied Economist" (see source below).
}
\source{
  Holden, Darryl and Roger Perman (1994), Unit Roots and Cointegration
  for the Economist, in:  Cointegration for the Applied Economist,
  ed. B. Bhaskara Rao, chapter 3, Data Appendix, Table D.3.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Rao Holden Perman}
  
