\name{Raotbl5} 
\alias{Raotbl5}
\encoding{latin1}
\docType{data}
\title{Data set used by Pierre Perron (1994)}
\description{
  This data set contains the time series used by Pierre Perron in his
  article: ''Trend, Unit Root and Structural Change in Macroeconomic
  Time Series". 
}
\usage{data(Raotbl5)}
\format{
  A data frame on real aggregate output for various countries; annual
  data starting in 1870 until 1986.
  
    \tabular{rl}{
    \code{ita} \tab Italy. \cr
    \code{nor} \tab Norway. \cr
    \code{swe} \tab Sweden. \cr
    \code{ukg} \tab United Kingdom. \cr
    \code{usa} \tab United States of America.
  }

  For further details about the data see \emph{Notes} in the data
  appendix `Table D.5' of Rao, ``Cointegration for the Applied
  Economist".

}
\source{
  Pierre Perron (1994), Trend, Unit Root and Structural Change in
  Macroeconomic Time Series, in:  Cointegration for the Applied Economist,
  ed. B. Bhaskara Rao, chapter 4, Data Appendix, Table D.5.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Rao Perron}
