\name{Raotbl6} 
\alias{Raotbl6}
\encoding{latin1}
\docType{data}
\title{Data set used by Yash P. Mehra (1994)}
\description{
  This data set contains quarterly data for the U.S.A. in Yash
  P. Mehra's article: ``Wage Growth and the Inflation Process: An
  Empirical Approach" for his wage-price equations. 
}
\usage{data(Raotbl6)}
\format{
  A data frame with quarterly data from 1959:1 until 1989:3. 
  
    \tabular{rl}{
    \code{rgnp} \tab Real GNP. \cr
    \code{pgnp} \tab Potential real GNP. \cr
    \code{ulc} \tab Unit labor cost. \cr
    \code{gdfco} \tab Fixed weight deflator for personal consumption
    expenditure excluding food and energy.\cr
    \code{gdf} \tab Fixed weight GNP deflator. \cr
    \code{gdfim} \tab Fixed weight import deflator. \cr
    \code{gdfcf} \tab Fixed weight deflator for food in personal consumption
    expenditure.  \cr
    \code{gdfce} \tab Fixed weight deflator for energy in personal
    consumption expenditure. 
  }

  Further details about the data can be found in the data appendix of
  Rao, ``Cointegration for the Applied Economist".
  
}
\source{
  Yash P. Mehra (1994), Wage Growth and the Inflation Process: An
  Empirical Approach, in:  Cointegration for the Applied Economist,
  ed. B. Bhaskara Rao, chapter 5, Data Appendix, Table D.6.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Rao Mehra}
