\name{Raotbl7} 
\alias{Raotbl7}
\docType{data}
\encoding{latin1}
\title{Data set used by Glenn Otto (1994)}
\description{
  This data set contains Canadian quarterly data used by Glenn Otto in his
  article: ``Diagnostic Testing: An Application to the Demand for M1". 
}
\usage{data(Raotbl7)}
\format{
  A data frame with quarterly data from 1956:1 until 1988:4. 
  
    \tabular{rl}{
    \code{m1} \tab Money stock M1. \cr
    \code{p} \tab Implicit price deflator for GDP, 1981=100. \cr
    \code{gdp} \tab GDP at constant 1981 prices. \cr
    \code{r} \tab 90-day prime corporate rate. 
  }
}
\source{
  Glenn Otto (1994), Diagnostic Testing: An Application to the Demand
  for M1, in:  Cointegration for the Applied Economist, 
  ed. B. Bhaskara Rao, chapter 6, Data Appendix, Table D.6.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set Rao Otto}
