\name{UKconinc}
\alias{UKconinc}
\docType{data}
\encoding{latin1}
\title{Data set for the United Kingdom}
\description{
  This data set contains the series used by Hylleberg, S., R. F. Engle,
  C. W. J. Granger and B. S. Yoo (1990), Seasonal Integration and
  Cointegration, Journal of econometrics, 44, 215--238.
}
\usage{data(UKconinc)}
\format{
  A data frame of quarterly data ranging from 1955:Q1
  until 1984:Q4. The data is expressed in natural logarithms.
 

    \tabular{rl}{
    \code{consl} \tab The log of total real consumption in the U.K. \cr
    \code{incl} \tab The log of real disposable income in the U.K.
    }
}
\source{
  Journal of Applied Econometrics Data Archive
  \url{http://qed.econ.queensu.ca/jae/}
}
\references{
Hylleberg, S., R. F. Engle, C. W. J. Granger and B. S. Yoo (1990),
Seasonal Integration and Cointegration, \emph{Journal of econometrics},
\bold{44}, 215--238.}
 
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set UK consumption income HEGY}
