\name{UKconsumption}
\alias{UKconsumption}
\docType{data}
\encoding{latin1}
\title{Data set for the United Kingdom}
\description{
  This data set contains the series used by in Charemza, W. (1997), New
  Directions in Econometric Practice, 2nd edition, Edward Elgar,
  Cheltenha, Uk. for analysing private in the United Kingdom.
}
\usage{data(UKconsumption)}
\format{
  A data frame of quarterly \code{ts} objects ranging from 1957:Q1
  until 1975:Q4.
 

    \tabular{rl}{
    \code{cons} \tab Consumers` non-durable expenditure in the U.K. in
    1970 prices. \cr
    \code{inc} \tab Personal disposable income in the U.K. in 1970
    prices. \cr
    \code{price} \tab Consumers` expenditure deflator index, 1970=100.
    }
}
\source{
  Pokorny, M. (1987), An Introduction to Econometrics, page 408, Basil
  Blackwell Ltd. Original data source: Economic Trends, Annual
  Supplements, 1976 and 1981, HMSO.
}
\references{
  Charemza, W. (1997), New Directions in Econometrics Practice, 2nd
  edition, Edward Elgar, Cheltenham, U.K.
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set UK consumption Charemza}
