\name{UKpppuip}
\alias{UKpppuip}
\docType{data}
\encoding{latin1}
\title{Data set for the United Kingdom: ppp and uip}
\description{
  This data set contains the series used by in Johansen and Juselius
  (1992), Testing structural hypothesis in a multivariate cointegration
  analysis of the PPP and UIP for UK, Journal of Econometrics, 53, 211-244.
}
\usage{data(UKpppuip)}
\format{
  A data frame of quarterly data ranging from 1971:Q1
  until 1987:Q2. All variables are expressed in logarithms.
 

    \tabular{rl}{
    \code{p1} \tab UK wholesale price index. \cr
    \code{p2} \tab Trade weighted foreign whole sale price index.\cr
    \code{e12} \tab UK effective exchange rate.\cr
    \code{i1} \tab Three-month treasury bill rate in the UK.\cr
    \code{i2} \tab Three-month Eurodollar interest rate.\cr
    \code{dpoil0} \tab World oil price at period \code{t}.\cr
    \code{dpoil1} \tab World oil price at period \code{t-1}.
  }
}
\references{
  Johansen, S. and K. Juselius (1992), Testing structural hypothesis in
  a multivariate cointegration analysis of the PPP and UIP for UK,
  Journal of Econometrics, 53, 211-244.
 
}
\author{Bernhard Pfaff}
\keyword{datasets}
\concept{data set UK PPP and UIP}
